[SW-23] refactor: PR 관련 요청사항 적용
