--- backend/src/main/java/com/done/swim/domain/pool/dto/requestDto/CreatePoolRequestDto.java
@@ -1,4 +1,0 @@
-package com.done.swim.domain.pool.dto.requestDto;
-
-public class CreatePoolRequestDto {
-}
