+++ backend/src/main/java/com/done/swim/domain/pool/dto/requestdto/CreatePoolRequestDto.java
@@ -1,0 +1,5 @@
+package com.done.swim.domain.pool.dto.requestdto;
+
+public class CreatePoolRequestDto {
+
+}
