--- backend/src/main/java/com/done/swim/domain/pool/dto/responseDto/CreatePoolResponseDto.java
@@ -1,4 +1,0 @@
-package com.done.swim.domain.pool.dto.responseDto;
-
-public class CreatePoolResponseDto {
-}
