+++ backend/src/main/java/com/done/swim/domain/pool/dto/responsedto/CreatePoolResponseDto.java
@@ -1,0 +1,5 @@
+package com.done.swim.domain.pool.dto.responsedto;
+
+public class CreatePoolResponseDto {
+
+}
