+++ backend/src/main/java/com/done/swim/domain/pool/dto/responsedto/PoolDetailResponseDto.java
@@ -1,8 +1,8 @@
-package com.done.swim.domain.pool.dto.responseDto;
+package com.done.swim.domain.pool.dto.responsedto;
 
 import com.done.swim.domain.pool.entity.Pool;
-import com.done.swim.domain.poolreview.dto.responseDto.PoolDetailPageReviewResponseDto;
-import com.done.swim.domain.swimmingtime.dto.FreeSwimmingTimeResponseDto;
+import com.done.swim.domain.poolreview.dto.responsedto.PoolDetailPageReviewResponseDto;
+import com.done.swim.domain.swimmingtime.dto.responsedto.FreeSwimmingTimeResponseDto;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
