+++ backend/src/main/java/com/done/swim/domain/pool/service/PoolService.java
@@ -1,6 +1,6 @@
 package com.done.swim.domain.pool.service;
 
-import com.done.swim.domain.pool.dto.responseDto.PoolDetailResponseDto;
+import com.done.swim.domain.pool.dto.responsedto.PoolDetailResponseDto;
 import com.done.swim.domain.pool.entity.Pool;
 import com.done.swim.domain.pool.repository.PoolRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,8 +16,8 @@ public class PoolService {
 
   public PoolDetailResponseDto poolsDetailPage(Long id) {
     Pool pool = poolRepository.findByIdWithCommentAndTimes(id)
+        //이 후 GlobalException 정의 후 수정 예정
         .orElseThrow(() -> new IllegalArgumentException("해당 수영장이 존재하지 않습니다."));
-
     return PoolDetailResponseDto.from(pool);
   }
 }
